requirements.txt
requests
jinja2
nose
py
gitflow
termcolor
argcomplete
tabulate
